+++ CHANGELOG.md
@@ -4,19 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.7.4]
+## [1.8.0]
 
 ### Added
 - New tile/touch action mushymato.MMAP_HoleWarpBuilding [X Y], warp that will put player into the building on this tile, for use with buildings.
 - New tile/touch action mushymato.MMAP_WarpBuildingOut [X Y], mushymato.MMAP_MagicWarpBuildingOut [X Y], mushymato.MMAP_HoleWarpBuildingOut [X Y], to add additional nonstandard exits from building interior.
 - New tile/touch action mushymato.MMAP_If GSQ ## if-case ## else-case
 - New tile/trigger action mushymato.MMAP_ToggleTAS X Y \<tasId\>+
+- New feature mushymato.MMAP/FloorPathProperties which allows paths to get tile properties.
 
 ### Changed
 - `mushymato.MMAP/DrawLayerRotate.` keys are renamed to `mushymato.MMAP/DrawLayer.`
 
 ### Fixed
 - Respawning temporary animated sprites now immediately spawn.
+- Light spawning at top left of tile instead of center.
 
 ## [1.7.3]